Add files via upload
It is an problem statement in LeetCode with Medium level of difficulty. It is based on graph theory. I have implemented the solution using python.
The basic approach is to use DFS(depth-first search) and traverse the nodes first . To find the group I have used an variable called visited which keeps track a new node which is not connected with previous nodes and needs a new traversal . The final result is stored in "g" variable which gets incremented when a new group needs to found.